src/requirements.txt
asgiref==3.6.0
autopep8==1.5.7
beautifulsoup4==4.11.1
behave==1.2.6
behave-django==1.4.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.1.5
django-cpf-cnpj==1.0.0
django-easy-audit==1.3.3
django-filter==21.1
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
django-cors-headers==4.3.1
drf-yasg==1.20.0
gunicorn==20.1.0
inflection==0.5.1
itypes==1.2.0
jwcrypto==1.0
Markdown==3.3.4
oauthlib==3.1.1
parse==1.19.0
parse-type==0.5.2
psycopg2-binary==2.9.10
pycodestyle==2.7.0
pyparsing==2.4.7
python-decouple==3.5
requests==2.32.4
rest-condition==1.0.3
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.17.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
toml==0.10.2
Unipath==1.1
uritemplate==4.1.1
urllib3==2.5.0
whitenoise==5.3.0
wrapt==1.17.2
celery>=5.3
redis>=4.0
django-celery-beat>=2.5
flower>=1.2.0
neo4j>=5.0
airbyte==0.27.0
airbyte-api==0.52.2
airbyte-cdk==6.56.7
airbyte_protocol_models_dataclasses==0.17.1
airbyte_protocol_models_pdv2==0.13.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
asn1crypto==1.5.1
attributes-doc==0.4.0
attrs==25.3.0
backoff==2.2.1
boltons==25.0.0
boto3==1.38.46
botocore==1.38.46
bracex==2.6
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==44.0.3
dataclasses-json==0.6.7
Deprecated==1.2.18
dpath==2.2.0
duckdb==1.3.1
duckdb_engine==0.13.6
dunamai==1.24.1
exceptiongroup==1.3.0
filelock==3.18.0
genson==1.3.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-bigquery==3.30.0
google-cloud-bigquery-storage==2.32.0
google-cloud-core==2.4.3
google-cloud-secret-manager==2.24.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
interchange==2021.0.4
isodate==0.6.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==0.2
jsonschema==4.17.3
langchain-core==0.1.42
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
monotonic==1.6
mypy_extensions==1.1.0
nltk==3.9.1
numpy==1.26.4
orjson==3.10.18
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pansi==2024.11.0
pillow==11.2.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==6.31.1
psutil==6.1.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
py2neo==2021.2.4
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGithub==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyrate-limiter==3.1.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ulid==3.0.0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests-cache==1.2.1
requests-toolbelt==1.0.0
rich==13.9.4
rich-click==1.8.9
rsa==4.9.1
s3transfer==0.13.0
serpyco-rs==1.15.2
sniffio==1.3.1
snowflake-connector-python==3.15.0
snowflake-sqlalchemy==1.7.5
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
sqlalchemy-bigquery==1.12.0
structlog==24.4.0
tenacity==8.5.0
tomlkit==0.13.3
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
url-normalize==2.2.1
uuid7==0.1.0
wcmatch==10.0
whenever==0.6.17
xmltodict==0.14.2